Change: Rename `compile` to `codegen`
