+++ src/cli/jit.rs
@@ -1,7 +1,7 @@
 use super::error::CLIError;
 use super::opts::RunOpt;
 use super::shell::Shell;
-use crate::compile::codegen;
+use crate::codegen::compile;
 use crate::parser;
 use crate::transform::TransformManager;
 use crate::translator::translate_ast;
@@ -17,7 +17,7 @@ use std::io::prelude::*;
 use std::process;
 
 pub fn run(opt: &RunOpt) -> Result<!, Error> {
-    codegen::initialize_native()?;
+    compile::initialize_native()?;
 
     if let Some(path) = &opt.input {
         let mut f = File::open(path).map_err(|_| CLIError::NotFound { path: path.clone() })?;
@@ -73,7 +73,7 @@ pub fn compile_jit(
         eprintln!("Transformed EIR:\n{}\n", transformed);
     }
 
-    let result = codegen::compile_eir(transformed, module_name)?;
+    let result = compile::compile_eir(transformed, module_name)?;
 
     if opt.print_ir {
         eprintln!("LLVM IR: \n{}", result.llvm_ir());
