+++ src/codegen/mod.rs
@@ -1,4 +1,4 @@
-pub mod codegen;
 pub mod compilation_result;
+pub mod compile;
 pub mod error;
 pub mod target_machine;
